pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["andrewqa1 <xieneuschannel@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.1.2"
pytest-cov = "^4.1.0"
coveralls = "^3.3.1"
pyyaml = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "code.scripts.generate_diff:main"


